requirement.txt
numpy
opencv-python
torch
icecream
tqdm
pyhocon
scipy
PyMCubes
trimesh
